requirements.txt
svgpathtools
Pillow
numpy